move user serializers to users
